Split resolver value function metadata tests
