+++ src/typechecker/tests/resolver_value_metadata/function_metadata/generic_parameters.rs
@@ -1,0 +1,122 @@
+use super::*;
+
+#[test]
+fn check_program_with_symbols_validates_resolver_function_type_parameter_counts() {
+    let program = parse_program(
+        r#"
+identity<T> = (value: T) T { value }
+"#,
+    );
+    let mut symbols = crate::resolver::Resolver::new()
+        .resolve_program(&program)
+        .expect("resolver succeeds");
+    symbols.set_type_parameter_count_for_test(Namespace::Value, "identity", Some(0));
+    let mut tc = TypeChecker::new();
+
+    let err = tc
+        .check_program_with_symbols(&program, &symbols)
+        .expect_err("resolver function generic arity mismatch should fail");
+
+    assert!(
+        err.iter().any(|d| d
+            .message
+            .contains("resolver value symbol 'identity' has type parameter count 0, expected 1")),
+        "expected resolver function generic arity diagnostic, got {err:?}"
+    );
+}
+
+#[test]
+fn check_program_with_symbols_validates_resolver_function_type_parameter_names() {
+    let program = parse_program(
+        r#"
+identity<T> = (value: T) T { value }
+"#,
+    );
+    let mut symbols = crate::resolver::Resolver::new()
+        .resolve_program(&program)
+        .expect("resolver succeeds");
+    symbols.set_type_parameter_names_for_test(
+        Namespace::Value,
+        "identity",
+        Some(vec!["U".to_string()]),
+    );
+    let mut tc = TypeChecker::new();
+
+    let err = tc
+        .check_program_with_symbols(&program, &symbols)
+        .expect_err("resolver function generic parameter name mismatch should fail");
+
+    assert!(
+        err.iter().any(|d| d.message.contains(
+            "resolver value symbol 'identity' has type parameter names '(U)', expected '(T)'"
+        )),
+        "expected resolver function generic parameter name diagnostic, got {err:?}"
+    );
+}
+
+#[test]
+fn check_program_with_symbols_validates_resolver_function_type_parameter_bounds() {
+    let program = parse_program(
+        r#"
+Json: behavior {
+    encode: (Self) StaticString
+}
+encode<T: Json> = (value: T) StaticString { "encoded" }
+"#,
+    );
+    let mut symbols = crate::resolver::Resolver::new()
+        .resolve_program(&program)
+        .expect("resolver succeeds");
+    symbols.set_type_parameter_bounds_for_test(
+        Namespace::Value,
+        "encode",
+        Some(vec![("T".to_string(), "Other".to_string())]),
+    );
+    let mut tc = TypeChecker::new();
+
+    let err = tc
+        .check_program_with_symbols(&program, &symbols)
+        .expect_err("resolver function generic bound mismatch should fail");
+
+    assert!(
+        err.iter().any(|d| d.message.contains(
+            "resolver value symbol 'encode' has type parameter bounds '(T: Other)', expected '(T: Json)'"
+        )),
+        "expected resolver function generic bound diagnostic, got {err:?}"
+    );
+}
+
+#[test]
+fn check_program_with_symbols_validates_resolver_function_type_parameter_bound_refs() {
+    let program = parse_program(
+        r#"
+Json<T>: behavior {
+    encode: (Self) T
+}
+identity<T: Json<T>> = (value: T) T { value }
+"#,
+    );
+    let mut symbols = crate::resolver::Resolver::new()
+        .resolve_program(&program)
+        .expect("resolver succeeds");
+    symbols.set_type_parameter_bound_refs_for_test(
+        Namespace::Value,
+        "identity",
+        Some(vec![TypeParameterBoundRefMetadata {
+            type_parameter: "T".to_string(),
+            behavior: "Json".to_string(),
+            type_args: vec![AstType::Str],
+        }]),
+    );
+    let mut tc = TypeChecker::new();
+
+    let err = tc
+        .check_program_with_symbols(&program, &symbols)
+        .expect_err("resolver function generic bound ref mismatch should fail");
+
+    let expected = "resolver value symbol 'identity' has type parameter bound refs '(T: Json<StaticString>)', expected '(T: Json<T>)'";
+    assert!(
+        err.iter().any(|d| d.message.contains(expected)),
+        "expected resolver function generic bound ref diagnostic, got {err:?}"
+    );
+}
